wip start moving value stuff
